Fix not running test with gradle.
Gradle did not run the test which were written with jmockit. Therefore all test are re-written to use mockito. That was necessary, because updating the jmockit dependency allow to run test with gradle but that broke many of the tests.

In order to test with mockito there was refactoring done, especially breadking functionality into service classes which increase the testability.

The version is increased to 0.1.0!