Use docker compose down command for container removal.
And minor improvements.

Increase the script-engine version to 0.2.0.

The docker compose stop and docker compose remove command were both too slow. So the ProActive node killed the thread during the removal process.

The docker compose down command is now utilised to remove and kill the container within one go, which increases the likelihood of the container actually being removed.

The DockerComposeScriptEngineFactory got its own ProcessBuilder service class.

The DockerComposeScriptEngineFactory got a standard constructor with which it can be initialized at during the node startup.